+++ src/main/java/disaster/pulse/api/repository/TipoEventoRepository.java
@@ -3,5 +3,5 @@ package disaster.pulse.api.repository;
 import disaster.pulse.api.model.TipoEvento;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TipoEventoRepository extends JpaRepository<TipoEvento, Integer> {
+public interface TipoEventoRepository extends JpaRepository<TipoEvento, Long> {
 }
